requirements.txt
openai==0.28.0
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1
pandas==2.0.3